Remove duplicate translations, relocate component
