+++ resources/assets/js/components/JobBuilder/WorkEnv/JobWorkEnv.tsx
@@ -7,6 +7,7 @@ import {
   IntlShape,
   MessageDescriptor,
 } from "react-intl";
+import { formMessages } from "./WorkEnvForm";
 
 export const physEnvMessages = defineMessages({
   openConcept: {
@@ -192,11 +193,7 @@ export const JobWorkEnv: React.FunctionComponent<JobWorkEnv &
           data-c-margin="top(half) bottom(half)"
           data-c-font-weight="bold"
         >
-          <FormattedMessage
-            id="jobBuilder.workEnvModal.teamSize"
-            defaultMessage="Team Size"
-            description="Title for Team size section."
-          />
+          <FormattedMessage {...formMessages.teamSizeLabel} />
         </span>
         <span data-c-margin="left(normal)">{teamSize}</span>
       </div>
@@ -206,11 +203,7 @@ export const JobWorkEnv: React.FunctionComponent<JobWorkEnv &
           data-c-margin="top(half) bottom(half)"
           data-c-font-weight="bold"
         >
-          <FormattedMessage
-            id="jobBuilder.workEnvModal.physicalEnvLabel"
-            defaultMessage="Our Physical Environment"
-            description="The label displayed on the physical environment checkbox group."
-          />
+          <FormattedMessage {...formMessages.physicalEnvLabel} />
         </p>
         <div data-c-margin="left(quarter)">
           <div data-c-grid="gutter">
@@ -240,11 +233,7 @@ export const JobWorkEnv: React.FunctionComponent<JobWorkEnv &
           data-c-padding="top(half) bottom(half)"
           data-c-font-weight="bold"
         >
-          <FormattedMessage
-            id="jobBuilder.workEnvModal.technologyLabel"
-            defaultMessage="Technology"
-            description="The label displayed on the technology checkbox group."
-          />
+          <FormattedMessage {...formMessages.technologyLabel} />
         </p>
         <div data-c-margin="left(quarter)">
           <div data-c-grid="gutter">
@@ -274,11 +263,7 @@ export const JobWorkEnv: React.FunctionComponent<JobWorkEnv &
           data-c-margin="top(half) bottom(half)"
           data-c-font-weight="bold"
         >
-          <FormattedMessage
-            id="jobBuilder.workEnvModal.amenitiesLabel"
-            defaultMessage="Amenities"
-            description="The label displayed on the amenities checkbox group."
-          />
+          <FormattedMessage {...formMessages.amenitiesLabel} />
         </p>
         <div data-c-margin="left(quarter)">
           <div data-c-grid="gutter">
@@ -308,11 +293,7 @@ export const JobWorkEnv: React.FunctionComponent<JobWorkEnv &
           data-c-colour="c1"
           data-c-font-weight="bold"
         >
-          <FormattedMessage
-            id="jobBuilder.workEnv.moreOnWorkEnv"
-            defaultMessage="More about your Environment"
-            description="The title for the more about your environment textbox."
-          />
+          <FormattedMessage {...formMessages.moreOnWorkEnv} />
         </p>
         <p>{envDescription}</p>
       </div>
