+++ resources/assets/js/translations/locales/fr.json
@@ -551,13 +551,9 @@
   "jobBuilder.workEnv.textAreaPlaceholder1": "Essayez d’adopter un ton décontracté, franc et amical.",
   "jobBuilder.workEnv.thisIsOptional": "C’est facultatif.",
   "jobBuilder.workEnv.title": "Environnement de travail",
-  "jobBuilder.workEnvModal.amenitiesLabel": "À proximité",
   "jobBuilder.workEnvModal.cancelLabel": "Retour en arrière",
   "jobBuilder.workEnvModal.confirmLabel": "Prochaine étape",
   "jobBuilder.workEnvModal.modalMiddleLabel": "Passer pour la révision",
-  "jobBuilder.workEnvModal.physicalEnvLabel": "Notre environnement physique",
-  "jobBuilder.workEnvModal.teamSize": "Taille de l’équipe",
-  "jobBuilder.workEnvModal.technologyLabel": "Technologie",
   "jobBuilder.workEnvModal.title": "Environnement de travail",
   "jobBuilder.workEnvModal.workCultureTitle": "Culture du travail",
   "jobCard.applicants": "{applicants, plural,\n                  =0 {No Applicants}\n                  one {# Applicant}\n                  other {# Applicants}\n                }",
